Add handler to filter unhandled mock requests
We only care if an unexpected call is made to api.github.com, as opposed
to the default of raising a console warning about everything.

The presence of a MSW console error/warning means that we are trying to
exercise GitHub API functionality that we haven't yet mocked.

Currently, we're raising a console error but as our confidence
increases, we can change this to throw an exception that will halt
execution.