Use test ID to identify Open IFC button
There is an open issue within the cypress-real-events library which
states that the `.realHover()` function is no longer compatible with
Google Chrome after version 100[1].

So we instead proceed with refactoring that test to use test IDs to
match the desired HTML element.

The way this failure shows up during testing is the first use of the
function in question works, but any subsequent usage by other test
cases within the same run will not work.

However, if you retry the test run, it will give a "false positive"
as long as you run within the same Cypress + Google Chrome process.

False positive is in quotes because if you run it again within the
same process pair, it works but if you execute on a different test
runner, you will get a failure.

An example of this consistent failure would be executing this test
suite in GitHub Actions where each run is handled by a new test
runner with a completely different process space.

1: https://github.com/dmtrKovalenko/cypress-real-events/issues/247